feat: adding more recordng and uploading functionality + sending to backend

+++ routes/api.py
@@ -2,14 +2,10 @@
 import logging
 from datetime import datetime
 from functools import wraps
-from typing import Dict, List
+from typing import List
 
 from flask import Blueprint, request, jsonify, g
-from langchain.chains.base import Chain
-from langchain.chains.llm import LLMChain
 from langchain.output_parsers import StructuredOutputParser, ResponseSchema
-from langchain_core.output_parsers import StrOutputParser
-from langchain_core.runnables import RunnablePassthrough
 from openai import OpenAI
 from pydub import AudioSegment
 from services.clinic import ClinicService
@@ -18,6 +14,9 @@ from services.vet import VetService
 from werkzeug.utils import secure_filename
 from langchain_openai import ChatOpenAI, OpenAIEmbeddings
 from langchain.prompts import ChatPromptTemplate
+import openai
+import os
+
 
 api_bp = Blueprint('api', __name__)
 
@@ -71,7 +70,7 @@ def me():
     return jsonify(g.user), 200
 
 
-ALLOWED_EXTENSIONS = {'mp3', 'mp4', 'wav', 'ogg'}
+ALLOWED_EXTENSIONS = {'mp3', 'mp4', 'wav', 'ogg', 'm4a'}
 
 
 def allowed_file(filename):
@@ -79,36 +78,57 @@ def allowed_file(filename):
 
 
 @api_bp.route('/upload-consultation', methods=['POST'])
+@authorize
 def upload_consultation():
+    logging.info("request.files: %s", request.files)
     file = get_uploaded_file(request)
     if file is None:
+        logging.warning("No file part in the request.")
         return jsonify({"error": "No file part"}), 400
 
     name = request.form.get('name', 'Unnamed Consultation')
     language = request.form.get('language', 'swedish')  # Default to Swedish if not specified
     file_path = save_temp_file(file)
-
+    logging.info("File saved to: %s", file_path)
     try:
+        logging.info("Processing file: %s", file_path)
         chunks = chunk_audio(file_path)
-        transcriptions = transcribe_chunks(chunks)
+        logging.info("Audio file chunked into %d parts.", len(chunks))
 
+        
+        transcriptions = transcribe_chunks(chunks)
         full_transcription = " ".join(transcriptions)
-        logging.info("Full transcription completed.")
+        logging.info("Full transcription completed. Length of transcription: %d characters", len(full_transcription))
 
         consultation = create_consultation_object(name, full_transcription, language)
         logging.info("Consultation object created: %s", consultation)
 
-        # Clean up the temporary file
-        os.remove(file_path)
+        # Get user info from the authorization token
+        user = g.user
+        clinic_id = user['user']['clinic_id']
+        vet_email = user['user']['email']
+        logging.info("Saving consultation for clinic_id: %s, vet_email: %s", clinic_id, vet_email)
 
+        # Save the consultation
+        consultation_id = VetService.save_consultation(clinic_id, vet_email, name, consultation)
+        logging.info("Consultation successfully saved to database.")
+
+        # Clean up the temporary file
+        if os.path.exists(file_path):
+            os.remove(file_path)
+            logging.info("Temporary file removed: %s", file_path)
+        else:
+            logging.warning("Temporary file not found for removal: %s", file_path)
+
+        consultation['id'] = consultation_id
         return jsonify(consultation), 200
     except Exception as e:
-        logging.error(f"Error processing consultation: {str(e)}")
+        logging.error("Error processing consultation. Exception: %s", str(e))
         # Clean up the temporary file in case of error
         if os.path.exists(file_path):
             os.remove(file_path)
+            logging.info("Temporary file removed after error: %s", file_path)
         return jsonify({"error": "Error processing consultation"}), 500
-
 
 def get_uploaded_file(request):
     if 'file' not in request.files:
@@ -123,7 +143,7 @@ def save_temp_file(file):
     return file_path
 
 
-def transcribe_chunks(chunks):
+def transcribe_chunks(chunks: List[AudioSegment]) -> List[str]:
     transcriptions = []
     for i, chunk in enumerate(chunks):
         chunk_path = os.path.join("/tmp", f"chunk_{i}.mp3")
@@ -134,8 +154,7 @@ def transcribe_chunks(chunks):
             transcriptions.append(transcription)
     return transcriptions
 
-
-def chunk_audio(file_path, chunk_length_ms=60000):
+def chunk_audio(file_path: str, chunk_length_ms: int = 60000) -> List[AudioSegment]:
     audio = AudioSegment.from_file(file_path)
     chunks = [audio[i:i + chunk_length_ms] for i in range(0, len(audio), chunk_length_ms)]
     return chunks
@@ -145,7 +164,6 @@ openai_client = OpenAI(
     # defaults to os.environ.get("OPENAI_API_KEY") 
     # api_key="My API Key",
 )
-
 
 def transcribe_audio(file):
     try:
@@ -160,34 +178,46 @@ def transcribe_audio(file):
         raise e
 
 
-import openai
-import os
+
 
 openai.api_key = os.getenv('OPENAI_API_KEY')
 
-# Define sections and subsections
-SECTIONS = {
-    "Allmän information": [
-        "Besöksorsak", "Anamnes", "Medicinsk historia", "Kostvanor",
-        "Vaccinationsstatus", "Tidigare operationer eller behandlingar"
-    ],
-    "Fysisk undersökning": [
-        "Temperatur", "Hjärtfrekvens", "Andningsfrekvens", "Vikt",
-        "Allmänt utseende", "Detaljerade undersökningsanteckningar"
-    ],
-    "Kliniska anteckningar": [
-        "Subjektiva observationer", "Objektiva observationer",
-        "Bedömning", "Plan"
-    ]
-}
-
-# Create response schemas for each subsection
+# Hardcoded response schemas with English section names mapped to Swedish
 RESPONSE_SCHEMAS = [
-    ResponseSchema(name=f"{section}_{subsection.replace(' ', '_')}",
-                   description=f"The content for {subsection} in {section}")
-    for section, subsections in SECTIONS.items()
-    for subsection in subsections
+    ResponseSchema(name="General_Information_Visit_Reason",
+                   description="The content for Visit Reason in General Information"),
+    ResponseSchema(name="General_Information_Anamnesis",
+                   description="The content for Anamnesis in General Information"),
+    ResponseSchema(name="General_Information_Medical_History",
+                   description="The content for Medical History in General Information"),
+    ResponseSchema(name="General_Information_Dietary_Habits",
+                   description="The content for Dietary Habits in General Information"),
+    ResponseSchema(name="General_Information_Vaccination_Status",
+                   description="The content for Vaccination Status in General Information"),
+    ResponseSchema(name="General_Information_Previous_Surgeries_or_Treatments",
+                   description="The content for Previous Surgeries or Treatments in General Information"),
+    ResponseSchema(name="Physical_Examination_Temperature",
+                   description="The content for Temperature in Physical Examination"),
+    ResponseSchema(name="Physical_Examination_Heart_Rate",
+                   description="The content for Heart Rate in Physical Examination"),
+    ResponseSchema(name="Physical_Examination_Respiratory_Rate",
+                   description="The content for Respiratory Rate in Physical Examination"),
+    ResponseSchema(name="Physical_Examination_Weight",
+                   description="The content for Weight in Physical Examination"),
+    ResponseSchema(name="Physical_Examination_General_Appearance",
+                   description="The content for General Appearance in Physical Examination"),
+    ResponseSchema(name="Physical_Examination_Detailed_Examination_Notes",
+                   description="The content for Detailed Examination Notes in Physical Examination"),
+    ResponseSchema(name="Clinical_Notes_Subjective_Observations",
+                   description="The content for Subjective Observations in Clinical Notes"),
+    ResponseSchema(name="Clinical_Notes_Objective_Observations",
+                   description="The content for Objective Observations in Clinical Notes"),
+    ResponseSchema(name="Clinical_Notes_Assessment",
+                   description="The content for Assessment in Clinical Notes"),
+    ResponseSchema(name="Clinical_Notes_Plan",
+                   description="The content for Plan in Clinical Notes")
 ]
+
 
 # Initialize the structured output parser
 PARSER = StructuredOutputParser.from_response_schemas(RESPONSE_SCHEMAS)
@@ -198,26 +228,26 @@ TEMPLATE = """You are an AI that excels at extracting key points from text. Plea
 Extract the following sections and their subsections from the given text. If a particular piece of information is not present, output "Not specified".
 
 Sections:
-1. Allmän information
-   - Besöksorsak
-   - Anamnes
-   - Medicinsk historia
-   - Kostvanor
-   - Vaccinationsstatus
-   - Tidigare operationer eller behandlingar
+1. General Information
+   - Visit Reason
+   - Anamnesis ()
+   - Medical History
+   - Dietary Habits
+   - Vaccination Status
+   - Previous Surgeries or Treatments
 
-2. Fysisk undersökning
-   - Temperatur
-   - Hjärtfrekvens
-   - Andningsfrekvens
-   - Vikt
-   - Allmänt utseende
-   - Detaljerade undersökningsanteckningar
+2. Physical Examination
+   - Temperature
+   - Heart Rate
+   - Respiratory Rate
+   - Weight
+   - General Appearance
+   - Detailed Examination Notes
 
-3. Kliniska anteckningar
-   - Subjektiva observationer
-   - Objektiva observationer
-   - Bedömning
+3. Clinical Notes
+   - Subjective Observations
+   - Objective Observations
+   - Assessment
    - Plan
 
 Text: {transcription}
@@ -230,23 +260,41 @@ Please provide your response in {language}.
 PROMPT = ChatPromptTemplate.from_template(TEMPLATE)
 
 # Initialize the language model
-LLM = ChatOpenAI(model_name="gpt-4", temperature=0)
+LLM = ChatOpenAI(model_name="gpt-4o-mini", temperature=0)
 
 
 # Function to organize the parsed output into structured sections
 def organize_sections(parsed_output: dict) -> dict:
     return {
-        section: {
-            subsection.replace('_', ' '): parsed_output.get(f"{section}_{subsection.replace(' ', '_')}",
+        "general_information": {
+            "visit_reason": parsed_output.get("General_Information_Visit_Reason", "Not specified"),
+            "anamnesis": parsed_output.get("General_Information_Anamnesis", "Not specified"),
+            "medical_history": parsed_output.get("General_Information_Medical_History", "Not specified"),
+            "dietary_habits": parsed_output.get("General_Information_Dietary_Habits", "Not specified"),
+            "vaccination_status": parsed_output.get("General_Information_Vaccination_Status", "Not specified"),
+            "previous_surgeries_or_treatments": parsed_output.get(
+                "General_Information_Previous_Surgeries_or_Treatments", "Not specified")
+        },
+        "physical_examination": {
+            "temperature": parsed_output.get("Physical_Examination_Temperature", "Not specified"),
+            "heart_rate": parsed_output.get("Physical_Examination_Heart_Rate", "Not specified"),
+            "respiratory_rate": parsed_output.get("Physical_Examination_Respiratory_Rate", "Not specified"),
+            "weight": parsed_output.get("Physical_Examination_Weight", "Not specified"),
+            "general_appearance": parsed_output.get("Physical_Examination_General_Appearance", "Not specified"),
+            "detailed_examination_notes": parsed_output.get("Physical_Examination_Detailed_Examination_Notes",
                                                             "Not specified")
-            for subsection in subsections
+        },
+        "clinical_notes": {
+            "subjective_observations": parsed_output.get("Clinical_Notes_Subjective_Observations", "Not specified"),
+            "objective_observations": parsed_output.get("Clinical_Notes_Objective_Observations", "Not specified"),
+            "assessment": parsed_output.get("Clinical_Notes_Assessment", "Not specified"),
+            "plan": parsed_output.get("Clinical_Notes_Plan", "Not specified")
         }
-        for section, subsections in SECTIONS.items()
     }
 
 
 # Create the extraction chain using LLMChain
-extraction_chain =  PROMPT | LLM | PARSER
+extraction_chain = PROMPT | LLM | PARSER
 
 
 def extract_sections(transcription: str, language: str = "swedish") -> dict:
@@ -254,7 +302,7 @@ def extract_sections(transcription: str, language: str = "swedish") -> dict:
     result = extraction_chain.invoke(
         {"transcription": transcription, "language": language, "format_instructions": PARSER.get_format_instructions()})
     logging.info("Extracted sections: %s", result)
-    return result
+    return organize_sections(result)
 
 
 def create_consultation_object(name: str, full_transcription: str, language: str = "swedish") -> dict:
